correction

+++ app/Http/Controllers/PhotoGallaryController.php
@@ -171,12 +171,15 @@ class PhotoGallaryController extends Controller
     }
 
     public function product_photo_delete(Request $request){
-        echo  $request->get('id');die;
-        DB::table('product_images')->where('id',$id)->delete();
-        $path = public_path('uploads/product_images/').$filename;
-        if (file_exists($path)) {
-            unlink($path);
+        DB::table('product_images')->where('id', $request->get('id'))->delete();die;
+        $data = DB::table('product_images')->where('id',$request->get('id'))->first();
+        if($data){
+            $path = public_path('uploads/product_images/').$data->image;
+            if (file_exists($path)) {
+                unlink($path);
+            }
+            DB::table('product_images')->where('id', $request->get('id'))->delete();
         }
-        return response()->json(['success'=>$filename]);
+        return response()->json(['success'=>$data->image]);
     }
 }
